Allow changing the layout of workspaces by storing it (Thanks mhcerri)
Before commit 4976fa33504907510f1ee9e44884c116af0d801b, setting the
layout of workspaces to something else than the default would just mess
up the parent container of the workspace (the content container).

After that commit, it would create an unnecessary split container when
you change the layout _before_ opening any containers. To avoid this, we
now store the layout (similar to how the 'workspace_layout'
configuration directive works) and apply it when the first container is
attached to the workspace.

Fixes #796